MyJobs and Results works
The “Results” and “My Jobs” pages are fully functional. They include the ability to save a preferred job from the “Results” page to the database and delete saved jobs from the “My Jobs” page. The pages are rendered using the JobCard component, which retrieves job data based on JobID from the database.

Additionally, Middleware is used to decode the token retrieved from cookies. All necessary routers are in place. I have also added the required routes in main.tsx on the client side to ensure proper page loading.